feat: add routes and reorganise
Add a route for manual-test, and reorganise the home page into a
HomeComponent.
