+++ public/src/app/home/home.component.css
@@ -1,0 +1,326 @@
+.flex-box {
+  flex-direction: column;
+  display: flex;
+  position: fixed;
+  left: 0;
+  top: 0;
+  right: 0;
+  bottom: 0;
+}
+
+.flex-container {
+  flex: 1;
+  overflow-x: auto;
+  overflow-y: scroll;
+}
+
+.navbar-phase {
+  background: rgb(240,240,240);
+  border-radius: 3px;
+  font-size: 1.2rem;
+  font-weight: bold;
+  padding: 4px 6px;
+  color: #343a40;
+  display: inline-block;
+  margin: 0 8px 0 0;
+}
+
+.navbar-phase a {
+  color: #343a40;
+}
+
+.navbar-phase-dates {
+  padding-left: 8px;
+  font-weight: normal;
+  color: #646464;
+}
+
+.navbar-new-pulls a.label {
+  color: white;
+  background: rgb(255, 24, 48);
+}
+
+.label {
+  font-size: 1rem;
+  padding: 4px;
+  margin: 0 4px;
+  border-radius: 4px;
+  display: inline-block;
+}
+
+.label.hide-if-zero {
+  display: none;
+}
+
+.label.hide-if-zero.failure {
+  display: inline-block;
+}
+
+.not-in-beta th#th-beta {
+  color: #606060;
+}
+
+.not-in-beta #tier-beta .tier-build-state,
+.not-in-beta #tier-beta ::ng-deep .tier-release-version a {
+  filter: brightness(1.25) saturate(0.25);
+}
+
+.tier app-sentry {
+  vertical-align: middle;
+  display: inline-block;
+  margin-left: 8px;
+}
+
+.emoji-group.emoji-group-border {
+  border: solid 1px #c0c0c0;
+}
+
+.emoji-group {
+  border: solid 1px #c0c0c000;
+  border-radius: 6px;
+  display: inline-block;
+  padding: 3px 0 4px 0;
+  margin: 1px 2px;
+}
+
+.emoji-group .emoji {
+  font-size: 16px;
+  font-family: "Noto Color Emoji";
+  vertical-align: middle;
+  display: inline-block;
+  width: 24px;
+  text-align: center;
+  margin-left: 4px;
+}
+
+/* Background of pull request and build status */
+
+.label {
+  background: #c0c0c0;
+  color: black;
+}
+
+a.label.success {
+  color: white;
+  background: rgb(68, 204, 17);
+}
+
+a.label.failure {
+  color: white;
+  background: rgb(255, 24, 48);
+}
+
+a.label.pending {
+  color: black;
+  background: rgb(240,240,100);
+}
+
+a.label.missing {
+  color: white;
+  background: rgb(12, 24, 48);
+}
+
+/* Tier Test Status */
+
+div.tier-test {
+  padding: 4px;
+  display: inline-block;
+  height: 32px;
+  vertical-align: top;
+  margin-left: 10px;
+  border-radius: 12px;
+  width: 26px;
+  font-size: 10pt;
+  text-align: center;
+  padding-top: 5px;
+}
+
+div.tier-test a,
+div.tier-test a:hover {
+  text-decoration: none;
+}
+
+div.tier-test-success {
+  color: white;
+  background-color: rgb(68, 204, 17);
+}
+
+div.tier-test-failure {
+  color: white;
+  background-color: rgb(255, 24, 48);
+}
+
+div.tier-test-pending {
+  color: black;
+  background: rgb(240,240,100);
+}
+
+/* Issue total */
+
+a.issue-count {
+  font-size: 9pt;
+  color: #808080;
+  margin-right: 8px;
+  text-decoration: none;
+}
+
+/* Etc */
+
+td.tier {
+  text-align: center;
+}
+
+td, th {
+  vertical-align: middle;
+}
+
+.tier-container {
+  display: inline-block;
+}
+
+.tier-release-date {
+  text-align: center;
+  font-size: 12px;
+}
+
+.tier-build-state a.label {
+  margin-right: 0;
+  border-radius: 4px 0 0 4px;
+  min-height: 32px;
+  max-height: 32px;
+  height: 32px;
+  padding-right: 8px;
+}
+
+.tier-build-state {
+  display: inline-block;
+  margin-left: 0px;
+  vertical-align: middle;
+}
+
+tbody th,
+td#tier-stable, td#tier-beta, td#tier-alpha {
+  white-space: nowrap;
+}
+
+thead th.tier {
+  text-align: center;
+}
+
+td.gap, th.gap {
+  width: 10%;
+}
+
+tr.tr-site td, tr.tr-site th {
+  text-align: center;
+  width: 9%;
+}
+
+tr.tr-site app-sentry {
+  margin-left: 8px;
+}
+
+td.sentry-events {
+  text-align: center;
+}
+
+/* Contributions */
+
+.navbar-contribution {
+  position: relative;
+  display: inline-block;
+}
+
+.navbar-contribution > span > img {
+  margin: -8px 4px 0;
+}
+
+.navbar-contribution-data {
+  display: block;
+  position: fixed;
+  background-color: #f9f9f9;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+  padding: 12px 16px;
+  z-index: 1;
+  left: 60px;
+  right: 60px;
+  visibility: hidden;
+  opacity: 0;
+  transition: visibility 0s, opacity 0.15s linear;
+  transition-delay: 0.25s;
+}
+
+.navbar-contribution:hover .navbar-contribution-data,
+.navbar-contribution.fixed .navbar-contribution-data {
+  visibility: visible;
+  opacity: 1;
+}
+
+.navbar-contribution.fixed .navbar-contribution-data {
+  z-index: 2;
+}
+
+.navbar-contribution.fixed > span > img {
+  border: solid 2px red;
+}
+
+.navbar-contribution-data thead th {
+  white-space: nowrap;
+  text-align: center;
+  vertical-align: middle;
+}
+
+.contribution-issue {
+  font-size: 12px;
+  display: inline-block;
+  background: #c0c0c0;
+  border-radius: 3px;
+  padding: 2px;
+  margin: 2px 4px;
+}
+
+table#sites-and-relatives {
+  border: none !important;
+}
+
+#sites-and-relatives .gap {
+  border: none;
+  background: white;
+  max-width: 0;
+  width: 0;
+  padding: 8px;
+  margin: 0;
+}
+
+#other-repos-cell {
+  position: relative;
+  cursor: default;
+}
+
+#other-repos {
+  display: none;
+  cursor: auto;
+  position: absolute;
+  bottom: 32px;
+  right: -1px;
+  z-index: 10;
+  background: white;
+  border: solid 1px #c0c0c0;
+  max-height: 640px;
+  overflow-y: auto;
+}
+
+#other-repos ul {
+  margin: 0.5em 1em;
+  padding-left: 1em;
+}
+
+#other-repos li {
+  font-size: 0.8em;
+  text-align: left;
+}
+
+#other-repos-cell:hover > #other-repos {
+  display: block;
+}
