+++ public/src/app/manual-test/manual-test.component.html
@@ -1,0 +1,1 @@
+<p>manual-test works!</p>
