+++ public/src/app/manual-test/manual-test.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-manual-test',
+  templateUrl: './manual-test.component.html',
+  styleUrls: ['./manual-test.component.css']
+})
+export class ManualTestComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
